Update to FreeType 2.1.5
svn path=/trunk/; revision=7814

+++ reactos/lib/freetype/src/smooth/ftgrays.c
@@ -4,7 +4,7 @@
 /*                                                                         */
 /*    A new `perfect' anti-aliasing renderer (body).                       */
 /*                                                                         */
-/*  Copyright 2000-2001, 2002 by                                           */
+/*  Copyright 2000-2001, 2002, 2003 by                                     */
 /*  David Turner, Robert Wilhelm, and Werner Lemberg.                      */
 /*                                                                         */
 /*  This file is part of the FreeType project, and may only be used,       */
@@ -340,7 +340,7 @@
                    long            byte_size )
   {
     ras.cells     = (PCell)buffer;
-    ras.max_cells = byte_size / sizeof ( TCell );
+    ras.max_cells = (int)( byte_size / sizeof ( TCell ) );
     ras.num_cells = 0;
     ras.area      = 0;
     ras.cover     = 0;
@@ -567,16 +567,16 @@
 
     if ( ex1 != ex2 )
     {
-      p     = ONE_PIXEL * ( y2 - y1 + delta );
-      lift  = (TCoord)( p / dx );
-      rem   = (TCoord)( p % dx );
+      p    = ONE_PIXEL * ( y2 - y1 + delta );
+      lift = (TCoord)( p / dx );
+      rem  = (TCoord)( p % dx );
       if ( rem < 0 )
       {
         lift--;
         rem += (TCoord)dx;
       }
 
-      mod -= dx;
+      mod -= (int)dx;
 
       while ( ex1 != ex2 )
       {
